Improve NFTs collections reload
- use LoadingDoggo loader
- refetch on demand - on displaying  the NFTs lists
