+++ ui/components/NFTS_update/NFTListCurrentWallet.tsx
@@ -4,12 +4,14 @@ import {
   selectCurrentAccountNFTsCount,
 } from "@tallyho/tally-background/redux-slices/selectors"
 import NFTList from "./NFTList"
-import { useBackgroundSelector } from "../../hooks"
+import { useBackgroundSelector, useNFTsReload } from "../../hooks"
 import NFTsExploreBanner from "./NFTsExploreBanner"
 
 export default function NFTListCurrentWallet(): ReactElement {
   const collections = useBackgroundSelector(selectCurrentAccountNFTs)
   const nftCount = useBackgroundSelector(selectCurrentAccountNFTsCount)
+
+  useNFTsReload()
 
   if (!nftCount) return <NFTsExploreBanner type="nfts" />
