+++ ui/hooks/nft-hooks.ts
@@ -10,7 +10,13 @@ import {
 } from "@tallyho/tally-background/redux-slices/utils/asset-utils"
 import { ETH } from "@tallyho/tally-background/constants"
 import { selectAssetPricePoint } from "@tallyho/tally-background/redux-slices/assets"
-import { useBackgroundSelector } from "./redux-hooks"
+import {
+  cleanCachedNFTs,
+  refetchCollections,
+  updateIsReloading,
+} from "@tallyho/tally-background/redux-slices/nfts_update"
+import { useEffect } from "react"
+import { useBackgroundDispatch, useBackgroundSelector } from "./redux-hooks"
 
 export const useTotalNFTsFloorPrice = (): {
   totalFloorPriceInETH: string
@@ -45,4 +51,16 @@ export const useTotalNFTsFloorPrice = (): {
     totalFloorPriceInETH: totalFloorPriceInETHFormatted,
     totalFloorPriceInUSD: totalFloorPriceLocalized,
   }
+}
+
+export const useNFTsReload = (): void => {
+  const dispatch = useBackgroundDispatch()
+
+  useEffect(() => {
+    dispatch(refetchCollections())
+    return () => {
+      dispatch(updateIsReloading(true)) // prepare for the next refetch
+      dispatch(cleanCachedNFTs())
+    }
+  }, [dispatch])
 }
