+++ ui/pages/NFTs.tsx
@@ -1,25 +1,20 @@
-import React, { ReactElement, useEffect, useState } from "react"
+import React, { ReactElement, useState } from "react"
 import classNames from "classnames"
 import {
   selectNFTBadgesCount,
   selectNFTsCount,
 } from "@tallyho/tally-background/redux-slices/selectors"
 import { useTranslation } from "react-i18next"
-import {
-  cleanCachedNFTs,
-  refetchCollections,
-} from "@tallyho/tally-background/redux-slices/nfts_update"
 import SharedPanelSwitcher from "../components/Shared/SharedPanelSwitcher"
 import NFTsExploreBanner from "../components/NFTS_update/NFTsExploreBanner"
 import NFTsHeader from "../components/NFTS_update/NFTsHeader"
-import { useBackgroundDispatch, useBackgroundSelector } from "../hooks"
+import { useBackgroundSelector, useNFTsReload } from "../hooks"
 import NFTListPortfolio from "../components/NFTS_update/NFTListPortfolio"
 import NFTListPortfolioBadges from "../components/NFTS_update/NFTListPortfolioBadges"
 
 const PANEL_NAMES = ["NFTs", "Badges"]
 
 export default function NFTs(): ReactElement {
-  const dispatch = useBackgroundDispatch()
   const nftCount = useBackgroundSelector(selectNFTsCount)
   const badgesCount = useBackgroundSelector(selectNFTBadgesCount)
   const { t } = useTranslation("translation", {
@@ -28,12 +23,7 @@ export default function NFTs(): ReactElement {
 
   const [panelNumber, setPanelNumber] = useState(0)
 
-  useEffect(() => {
-    dispatch(refetchCollections())
-    return () => {
-      dispatch(cleanCachedNFTs())
-    }
-  }, [dispatch])
+  useNFTsReload()
 
   return (
     <div className="page_content">
